Various band aids for ndsize_t → size_t issues
This is only relevant on 32bit system. All of
these belong to issue #473

+++ src/hdf5/BaseTagHDF5.cpp
@@ -132,7 +132,8 @@ void BaseTagHDF5::references(const std::vector<DataArray> &refs_new) {
     // extract vectors of names from vectors of new & old references
     std::vector<std::string> names_new(refs_new.size());
     transform(refs_new.begin(), refs_new.end(), names_new.begin(), util::toName<DataArray>);
-    std::vector<DataArray> refs_old(referenceCount());
+	//FIXME: issue 473
+    std::vector<DataArray> refs_old(static_cast<size_t>(referenceCount()));
     for (size_t i = 0; i < refs_old.size(); i++) refs_old[i] = getReference(i);
     std::vector<std::string> names_old(refs_old.size());
     transform(refs_old.begin(), refs_old.end(), names_old.begin(), util::toName<DataArray>);
